requirements.txt
openai
python-dotenv
PyMuPDF